requirements.txt
pkg-resources==0.0.0
psutil==5.6.1
pyserial==3.4
